Implement read+write+expire requesters in Redis
Now we save the requester's names to a set in redis.  Being a set, many
requests from the same user will only be counted once.  If there's only
one person, you get their name, otherwise you get the number of people.

Additionally, we set it to expire after a minute, so that if nobody
responds in a minute then your "coffee request" is cancelled and the
next request starts counting names from scratch.