Refactor to easily allow testing without YoAll
By default the app sends the resulting string to all subscribed Yo users
which can be pretty annoying for testing.  I wanted a test mode which
sends only to a test user (me) for quickly trying out new things.

I've refactored building the request in a DRY way so that you can turn
on test mode by setting an environment variable RUN_MODE to TEST and
setting TEST_USER to the target user.  This all happens in a small if at
the beginning of the file where I've also extracted some constants like
URLs to call, so that the actual request at the end is just one line
that pulls it all together.