Fall back to "nobody" if we couldn't get a requester name
This happens when there is a request but the number of names returned
from Redis is less than 1, i.e. 0.  In production this never happens
because Yo sends the user name with the request, so there will always be
at least one, however, in testing I sometimes omit the user name from
the URL, in which case nothing is inserted into Redis.  So, in case this
becomes an edge case in production, it's much better to have "nobody
wants a coffee" than " wants a coffee".